Updated usage of ImageCleaner
DatasetFormatter.from_toplosses removed the `ds_type` parameter.
In order to clean the whole dataset we need to create a separate
one with all the images. This commit alters the notebook
to reflect the new approach.

Some students, including myself, ran into an issue when ImageCleaner
didn't render at all. This happens when using Jupyter Lab instead
of Jupyter Notebook. I've added the warning to the notebook as well.

It also fixes an issue with creating ImageDataBunch from `cleaned.csv`.
The first param was `"."`, the current folder, but should be `path`.
This commit fixes it.