Update zone configuration for gcloud setup
The nvidia-tesla-p100 accelerator is not available in the gcloud us-west2-b zone, but is available in the us-west1-b zone (see https://cloud.google.com/compute/docs/gpus/#gpus-list). Not sure when this changed. Therefore, it appears that the us-west1-b zone should work for both regular and budget set-up options.