Update start_gcp.md
For some reason when I ran this and used the zone:

us-west1-b

I got the following error:

ERROR: (gcloud.compute.instances.create) Could not fetch resource:
 - The resource 'projects/static-protocol-256618/zones/us-west1-b/acceleratorTypes/nvidia-tesla-p4' was not found

So when I changed the zone to:

us-west2-c

It seemed to work ok so for some reason that instance type wasn't available in that zone when I tried today! So thats why suggested the change.