Add introductory posts
These lectures were my first introduction to Swift for TensorFlow and since taking them in the summer I've learned a lot more about the framework. I would like to do something in return and share two posts that might be helpful to beginners: 

- [Your first Swift for TensorFlow model](https://rickwierenga.com/blog/s4tf/s4tf-mnist.html)
- [An introduction to Generative Adversarial Networks in Swift for TensorFlow](https://rickwierenga.com/blog/s4tf/s4tf-gan.html)

Please let me know if this isn’t the right place for this, or if there is a better way to share my experience with the FastAI community. I'd love to do something to give back.

Thanks!