Update GCP docs to usable instances and zone (#503)
The current documentation describes instance and zone combinations as well as a GPU instance that are incompatible. The suggested instances are as close to the original as possible. The prices have also been updated to more accurately reflect the cost of running this. Finally the conda command needed to be updated, as it was giving an error using the current code. This is well documented in the forum (https://forums.fast.ai/t/platform-gcp/27375/703) and as an issue (https://github.com/fastai/course-v3/issues/493).