add instructions for updating requirements.txt
Users currently have to read through the fast.ai forum Render thread if they run into an error deploying because of the package versions used in the requirements.txt file. Given that this seems to be a very common issue, I suggest that this instruction be added to the documentation.